+++ Woof/Woof/View/OwnerFlow/BookingSlotView/BookingSlotView.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+/// A view presents the information related to booking process.
+struct BookingSlotView: View {
+    var body: some View {
+        Text(walkingsSlotStubText)
+    }
+
+    // MARK: - Private interface
+
+    private let walkingsSlotStubText = "Walkings slot stub"
+}
+
+#Preview {
+    BookingSlotView()
+}