+++ Woof/Woof/View/OwnerFlow/DetailedPetSitterView/DetailedPetSitterView.swift
@@ -6,50 +6,75 @@ struct DetailPetSitterView: View {
     @ObservedObject var viewModel: DetailSitterViewModel
 
     var body: some View {
-        VStack {
-            AvatarView(url: viewModel.imageURL)
-            FiveStarRatingView(stars: viewModel.rating)
-            TextWithIconLabelView(
-                iconName: .IconName.filledPerson,
-                text: viewModel.fullName
-            )
-            .contextMenu {
-                Button {
-                    viewModel.copyToClipboardText(viewModel.fullName)
-                } label: {
-                    CopyToClipboardLabel()
+        ScrollView {
+            HStack {
+                AvatarView(url: viewModel.imageURL)
+                SitterSectionInformationView(
+                    fullName: viewModel.fullName,
+                    rating: viewModel.rating,
+                    phoneNumber: viewModel.phoneNumber,
+                    city: viewModel.city
+                ) { text in
+                    viewModel.copyToClipboardText(text)
                 }
+                Spacer()
+                PriceLabel(price: viewModel.price)
             }
 
-            TextWithIconLabelView(
-                iconName: .IconName.phone,
-                text: viewModel.phoneNumber
-            )
-            .contextMenu {
-                Button {
-                    viewModel.copyToClipboardText(viewModel.phoneNumber)
-                } label: {
-                    CopyToClipboardLabel()
-                }
-            }
-
-            TextWithIconLabelView(
-                iconName: .IconName.house,
-                text: viewModel.city
-            )
-
+            Divider()
             Text(viewModel.bio)
-            Spacer()
+                .font(
+                    .system(
+                        size: AppStyle.FontStyle.footnote.size
+                    )
+                )
+            Divider()
+                .padding(.bottom)
+
+            HStack {
+                Text(slotHeaderLabel)
+                Spacer()
+                DatePicker("",
+                           selection: $viewModel.selectedDate,
+                           in: Date.now...,
+                           displayedComponents: .date)
+                    .labelsHidden()
+                    .font(
+                        .system(
+                            size: AppStyle.FontStyle.heading.size
+                        )
+                    )
+                    .tint(Color.App.purpleDark)
+                Spacer()
+                Button(searchButtonText) {
+                    Task {
+                        await viewModel.fetchWalkings()
+                    }
+                }
+                .buttonStyle(CapsuleWithWhiteText())
+            }
+            .padding()
+
+            if viewModel.isLoading {
+                ProgressView()
+            } else {
+                WalkingSlotsView(walkings: $viewModel.walkings)
+            }
         }
         .padding()
         .background(Color.App.grayLight)
     }
 
-    struct DetailPetSitterView_Previews: PreviewProvider {
-        static var previews: some View {
-            DetailPetSitterView(
-                viewModel: DetailSitterViewModel(sitter: Sitter.Dummy.emilyDoe)
-            )
-        }
+    // MARK: - Private interface
+
+    private let slotHeaderLabel = "Available slots"
+    private let searchButtonText = "Search"
+}
+
+struct DetailPetSitterView_Previews: PreviewProvider {
+    static var previews: some View {
+        DetailPetSitterView(
+            viewModel: DetailSitterViewModel(sitter: Sitter.Dummy.emilyDoe)
+        )
     }
 }